Move rule test fixtures to test/fixtures
