+++ test/core/rules/rule_test_cases_test.py
@@ -11,7 +11,7 @@ from sqlfluff.core.config import FluffConfig
 
 ids, test_cases = load_test_cases(
     test_cases_path=os.path.join(
-        os.path.abspath(os.path.dirname(__file__)), "test_cases", "*.yml"
+        "test/fixtures/rules/std_rule_cases", "*.yml"
     )
 )
