+++ test/core/rules/std_test.py
@@ -16,8 +16,8 @@ from test.fixtures.dbt.templater import (  # noqa
     in_dbt_project_dir,
     dbt_templater,
 )
-from test.fixtures.rules.L000 import Rule_L000
-from test.fixtures.rules.S000 import Rule_S000
+from test.fixtures.rules.custom.L000 import Rule_L000
+from test.fixtures.rules.custom.S000 import Rule_S000
 from sqlfluff.core.rules.std import get_rules_from_path
 
 
@@ -259,13 +259,13 @@ def test_rule_exception_is_caught_to_validation():
 def test_std_rule_import_fail_bad_naming():
     """Check that rule import from file works."""
     assert get_rules_from_path(
-        rules_path="test/fixtures/rules/*.py", base_module="test.fixtures.rules"
+        rules_path="test/fixtures/rules/custom/*.py", base_module="test.fixtures.rules.custom"
     ) == [Rule_L000, Rule_S000]
 
     with pytest.raises(AttributeError) as e:
         get_rules_from_path(
-            rules_path="test/fixtures/rules/bad_rule_name/*.py",
-            base_module="test.fixtures.rules.bad_rule_name",
+            rules_path="test/fixtures/rules/custom/bad_rule_name/*.py",
+            base_module="test.fixtures.rules.custom.bad_rule_name",
         )
 
     e.match("Rule classes must be named in the format of")
