+++ test/fixtures/rules/custom/S000.py
