+++ test/fixtures/rules/custom/bad_rule_name/E000.py
