+++ test/fixtures/rules/std_rule_cases/L001.yml
