+++ test/fixtures/rules/std_rule_cases/L002.yml
