+++ test/fixtures/rules/std_rule_cases/L003.yml
