+++ test/fixtures/rules/std_rule_cases/L005.yml
