+++ test/fixtures/rules/std_rule_cases/L006.yml
