+++ test/fixtures/rules/std_rule_cases/L010.yml
