+++ test/fixtures/rules/std_rule_cases/L018.yml
