+++ test/fixtures/rules/std_rule_cases/L019.yml
