+++ test/fixtures/rules/std_rule_cases/L022.yml
