+++ test/fixtures/rules/std_rule_cases/L023.yml
