+++ test/fixtures/rules/std_rule_cases/L024.yml
