+++ test/fixtures/rules/std_rule_cases/L025.yml
