+++ test/fixtures/rules/std_rule_cases/L026.yml
