+++ test/fixtures/rules/std_rule_cases/L027.yml
