+++ test/fixtures/rules/std_rule_cases/L032.yml
