+++ test/fixtures/rules/std_rule_cases/L039.yml
