+++ test/fixtures/rules/std_rule_cases/L040.yml
