+++ test/fixtures/rules/std_rule_cases/L042.yml
