+++ test/fixtures/rules/std_rule_cases/L043.yml
