+++ test/fixtures/rules/std_rule_cases/L045.yml
